requirements.txt
streamlit==1.38.0
supabase==2.7.1
python-dotenv==1.0.1
pandas==2.2.2
altair==5.4.1
reportlab==4.2.2
twilio==9.2.3
speechrecognition==3.10.4
whisper==20231117
openai==1.42.0
pyyaml==6.0.2
ratelimit==2.2.1
pytest==8.3.2